Streamline code, fix bugs, add tests, improve what AI did

+++ backend/WeightTracker.Api/Models/UserSettings.cs
@@ -3,8 +3,8 @@ namespace WeightTracker.Api.Models;
 public class UserSettings
 {
     public int Id { get; set; } = 1;
-    public decimal? HeightCm { get; set; }
+    public int? HeightCm { get; set; }
     public string PreferredUnit { get; set; } = "kg";
-    public decimal? TdeeKcal { get; set; }
+    public int? TdeeKcal { get; set; }
 
 }
